backend/requirements.txt
fastapi
unicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
requests
httpx